# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API of Julia 2",
        "title": "Julia API 2",
        "contact": {
            "name": "API Support",
            "url": "https://spotec.app/contato/",
            "email": "contact@spotec.app"
        },
        "version": "0.7.4"
    },
    "host": "api.julia.spotec.app",
    "basePath": "/api",
    "paths": {
...
        "/sample_models": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "add by json sample_model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample_models"
                ],
                "summary": "Add an sample_model",
                "parameters": [
                    {
                        "description": "Add sample_model",
                        "name": "sample_model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateSampleModelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SampleModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    }
                }
            }
        },
        "/sample_models/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get sample_model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample_models"
                ],
                "summary": "Get sample_model",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SampleModel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.SampleModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update by json sample_model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample_models"
                ],
                "summary": "Update an sample_model",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update sample_model",
                        "name": "sample_model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateSampleModelInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SampleModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete by sample_model ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample_models"
                ],
                "summary": "Delete an sample_model",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SampleModel"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    }
                }
            }
        },
        "/sample_models_select": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Select samples",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample_models"
                ],
                "summary": "Select samples",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by sample_string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit the number of results",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset the results",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SampleModel"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    }
                }
            }
        },
        "/samples": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get samples",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample_models"
                ],
                "summary": "Get samples",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by id or sample_string or sample_unique or sample_nullable",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.SampleModel"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/models.APIMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.APIMessage": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.SampleDetail": {
            "type": "object",
            "required": [
                "id",
                "sample_string"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "sample_string": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SampleItem": {
            "type": "object",
            "required": [
                "id",
                "sample_model_id",
                "sample_string"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "sample_model_id": {
                    "type": "integer"
                },
                "sample_string": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SampleModel": {
            "type": "object",
            "required": [
                "id",
                "sample_date",
                "sample_detail_id",
                "sample_double",
                "sample_string",
                "sample_unique"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "sample_date": {
                    "type": "string"
                },
                "sample_detail": {
                    "$ref": "#/definitions/models.SampleDetail"
                },
                "sample_detail_id": {
                    "type": "integer"
                },
                "sample_double": {
                    "type": "number"
                },
                "sample_items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SampleItem"
                    }
                },
                "sample_nullable": {
                    "type": "string"
                },
                "sample_string": {
                    "type": "string"
                },
                "sample_unique": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "requests.CreateSampleModelInput": {
            "type": "object",
            "required": [
                "sample_date",
                "sample_detail_id",
                "sample_double",
                "sample_string",
                "sample_unique"
            ],
            "properties": {
                "sample_date": {
                    "type": "string"
                },
                "sample_detail_id": {
                    "type": "integer"
                },
                "sample_double": {
                    "type": "number"
                },
                "sample_nullable": {
                    "type": "string"
                },
                "sample_string": {
                    "type": "string"
                },
                "sample_unique": {
                    "type": "string"
                }
            }
        },
...
        "requests.UpdateSampleModelInput": {
            "type": "object",
            "required": [
                "sample_date",
                "sample_detail_id",
                "sample_double",
                "sample_string",
                "sample_unique"
            ],
            "properties": {
                "sample_date": {
                    "type": "string"
                },
                "sample_detail_id": {
                    "type": "integer"
                },
                "sample_double": {
                    "type": "number"
                },
                "sample_nullable": {
                    "type": "string"
                },
                "sample_string": {
                    "type": "string"
                },
                "sample_unique": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}